requirements.txt
fake-bpy-module-4.2==20250130
